Extract API functions from admin-ui package (#576)
## Issue
https://github.com/jsdrupal/drupal-admin-ui/issues/571

The most basic set of helper functions extracted from the `admin-ui` package.
