+++ examples/admin_ui_widget_example/js/src/hello_world.route.js
@@ -1,5 +1,26 @@
-import React from 'react';
+import React, { Component, Fragment } from 'react';
+import { api } from '@drupal/admin-ui-utilities';
 
-const HelloWorld = () => <p>hello world</p>;
+class HelloWorld extends Component {
+  state = { routes: false };
+
+  componentDidMount() {
+    api('admin_ui_routes').then(({ routes }) => this.setState({ routes }));
+  }
+
+  render() {
+    return this.state.routes ? (
+      <Fragment>
+        {Object.values(this.state.routes).map(({ route, moduleName }) => (
+          <p key={route}>
+            {route} : {moduleName}
+          </p>
+        ))}
+      </Fragment>
+    ) : (
+      <p>routes loading</p>
+    );
+  }
+}
 
 export default HelloWorld;
