+++ examples/admin_ui_widget_example/package.json
@@ -3,7 +3,13 @@
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
+  "scripts": {
+    "build": "extension-points-bundler --mode=production"
+  },
   "dependencies": {
+    "@drupal/admin-ui-utilities": "file:../../packages/utilities"
+  },
+  "devDependencies": {
     "@drupal/extension-points-bundler": "file:../../packages/extension-points"
   }
 }
