+++ packages/utilities/src/index.js
@@ -1,0 +1,4 @@
+import api from './api';
+import { ApiError } from './errors';
+
+export { api, ApiError };
